+++ R/trajectory-activities.R
@@ -168,21 +168,28 @@ timeout.trajectory <- function(.trj, task) {
 #' @inheritParams set_attribute
 #' @param key the attribute name, or a callable object (a function) which
 #' must return the attribute name.
+#' @seealso \code{\link{set_attribute}}, \code{\link{set_global}}.
 #' @export
-timeout_from_attribute <- function(.trj, key, global=FALSE)
-  UseMethod("timeout_from_attribute")
+timeout_from_attribute <- function(.trj, key, global=FALSE) {
+  if (global) {
+    .Deprecated("timeout_from_global", old="timeout_from_attribute(global=TRUE)")
+    timeout_from_global(.trj, key)
+  }
+  else UseMethod("timeout_from_attribute")
+}
 
 #' @export
 timeout_from_attribute.trajectory <- function(.trj, key, global=FALSE) {
   check_args(key="string", global="flag")
-  .trj$timeout(key, global)
+  .trj$timeout(key, FALSE)
 }
 
 #' @rdname timeout
-#' @details \code{timeout_from_global} is a shortcut for
-#' \code{timeout_from_attribute(global=TRUE)}.
 #' @export
-timeout_from_global <- function(.trj, key) timeout_from_attribute(.trj, key, TRUE)
+timeout_from_global <- function(.trj, key) UseMethod("timeout_from_global")
+
+#' @export
+timeout_from_global <- function(.trj, key) .trj$timeout(key, TRUE)
 
 #' Set Attributes
 #'
@@ -193,26 +200,35 @@ timeout_from_global <- function(.trj, key) timeout_from_attribute(.trj, key, TRU
 #' must return attribute name(s).
 #' @param values numeric value(s) to set, or a callable object (a function) which
 #' must return numeric value(s).
-#' @param global if \code{TRUE}, the attribute will be global instead of per-arrival.
+#' @param global \code{global=TRUE} is deprecated. Use \code{*_global} instead.
 #' @param mod if set, \code{values} modify the attributes rather than substituting them.
 #' @param init initial value, applied if \code{mod} is set and the attribute was
 #' not previously initialised. Useful for counters or indexes.
 #'
 #' @return Returns the trajectory object.
-#' @seealso \code{\link{get_attribute}}.
+#' @seealso \code{\link{get_attribute}}, \code{\link{get_global}},
+#' \code{\link{timeout_from_attribute}}, \code{\link{timeout_from_global}}.
 #' @export
-set_attribute <- function(.trj, keys, values, global=FALSE, mod=c(NA, "+", "*"), init=0)
-  UseMethod("set_attribute")
+set_attribute <- function(.trj, keys, values, global=FALSE, mod=c(NA, "+", "*"), init=0) {
+  if (global) {
+    .Deprecated("set_global", old="set_attribute(global=TRUE)")
+    set_global(.trj, keys, values, mod, init)
+  }
+  else UseMethod("set_attribute")
+}
 
 #' @export
 set_attribute.trajectory <- function(.trj, keys, values, global=FALSE, mod=c(NA, "+", "*"), init=0)
-  .trj$set_attribute(keys, values, global, mod, init)
+  .trj$set_attribute(keys, values, mod, init)
 
 #' @rdname set_attribute
-#' @details \code{set_global} is a shortcut for \code{set_attribute(global=TRUE)}.
 #' @export
 set_global <- function(.trj, keys, values, mod=c(NA, "+", "*"), init=0)
-  set_attribute(.trj, keys, values, TRUE, mod, init)
+  UseMethod("set_global")
+
+#' @export
+set_global.trajectory <- function(.trj, keys, values, mod=c(NA, "+", "*"), init=0)
+  .trj$set_global(keys, values, mod, init)
 
 #' Activate/Deactivate Sources
 #'